Build/Gradle/Jandex/Quarkus: Attempt to fix the occasional `ConcurrentModificiationException`
The occasional `ConcurrentModificationException`s seem to always happen with or are related to the Jandex and Quarkus plugins. This change completely disables the Jandex tasks for non-`main` source sets, hoping that this reduces the probability to hit this error. The CME happens with all Gradle versions, it just seems to happen more often with newer Gradle versions.