CI: Disable tests in Windows + macOS jobs (#7403)
Running tests on GH Windows+Mac runners was always brittle and recently never succeeded. The value of these tests is quite low, so CI for these environments is now limited to building Nessie and running only unit but no integration tests on macOS and no tests on Windows GH runners.