Refactor `ContentMapping.fetchContents`
There is a *very* low chance (only for old, migrated repositories) that the exact same content object is used for multiple content-keys. If that situation happens, restart with a special, slower and more expensive implementation. It could have only happened before the enforcement of new content-IDs for Put-operations was in place _and_ if the content-ID was explicitly specified by a client.