Avoid re-running previously validated access checks on commit retries.
This is mainly to speed up the retry cycle so as to improve
throughput and reduce individual commit times.

Authorization is not dependent on commit history.

Races between authorization data changes and commits were possible
even before this change, because there are no guarantees as to the
relative timing of AuthZ checks (even without retries) and finilizing
the commit.