Store-cache: allow cache sizing based on max-heap-size (#7540)
This changes the way Nessie's object cache is configured. The previous default was to specify a fixed size in MB. When dealing with different JVM configurations / pods, it's easier to have a reasonable default of, as in this change, 70% of the Java max heap setting. This allows to have a single configuration that works in pretty much all situations/configurations.

If a fixed size is already configured, consider removing that explicit value and use fractional cache size setting based on the JVM's max heap size.