Storage: record previous HEADs of named references
Persisted `Reference`s now keep a history of previous HEADs, limited in size and age, defaults to up to 20 elementa with a maximum age of 300 seconds.

Exposes a new Nessie-API V2 endpoint to retrieve the status of a reference. The returned information contains a consistency status of the HEAD commit, the recorded recent HEADs and, if requested, a combined status for the recent commits in the commit log.

This change also ensures that a "broken" reference can be re-assigned to another existing commit, and that it can be deleted.