BigTable: reimplement scanAllObjects with server-side filtering
This commit changes the BigTable implementation of scanAllObjects
to use server-side instead of client-side filtering on object types.

This is implemented using a condition filter. Condition filters are
generally not recommended because they are slower, but scanAllObjects
is not meant to be particularly efficient. The fact that we are also
limiting the query to a row prefix should alleviate the performance
impact.