ITOAuth2Client: improve shutdown sequence
Remarked while working on #7673: the call to close()
may disrupt ongoing token refreshes and / or refresh
schedule attempts; the client does close normally
eventually, but if DEBUG logs are enabled, many
strange errors may be logged. Also, if the executor
is scheduling a refresh when close() is called, this
may trigger the start of a thread, only to see it
being closed immediately after.

By guarding the shutdown sequence with a boolean flag,
we can achieve a much smoother shutdown sequence.