Quarkus-tests: do not log OIDC connection + tracing warnings
The "well known OIDC connection error" that is logged at startup (during automatic OIDC endpoint discovery) cannot be disabled - also OIDC cannot be disabled at runtime (it's a build-time config setting). Can only turn off logging for the class that emits the warning - but should only be done in tests, but not in production. This change is to reduce "log spam" in tests.

Also configures trace-sampling to default to 'none' for tests.