Custom obj types should not persist class name (#7915)
Custom object types do not need to and should not persist the fully qualified _implementation_ class for the custom object type, because it can be inferred from the type. This saves a bunch of storage but more importantly allows later refactorings of custom obj types. Since custom object types are not used yet, this change is safe. Not changing the column names for CQL + SQL in case users have already migrated their tables.