Disable OpenTelemetry SDK when no endpoint is defined (#8004)
This change introduces a conditional enablement of OpenTelemetry SDK,
depending on whether a collector endpoint has been explicitly provided
by the user or not.

A new config source interceptor detects whether the collector endpoint
has been user-defined or not. If it was, then leave OpenTelemetry enabled; if
not, disable OpenTelemetry.