CI/Testing: Let Renovate manage Docker image references (#8015)
Use specific versions for Docker images instead of `latest`, versions to be managed by Renovate.

Adds a "dummy" `Dockerfile` for each referenced image, which can be managed by Renovate. The code that uses the Docker image extracts the image name and tag from that `Dockerfile`.

The code to extract the image name and tag is duplicated in a couple of places, but introducing a separate project just for this functionality feels overkill. It is rather unlikely that the image/tag parsing code will change. Also some "dummy" `Dockerfile`s are duplicated.

Fixes #8014