Add Apache HTTP client, choose HTTP client by name
This change adds the Apache HTTP client (synchronous, HTTP/1.1 only for now).

The Apache HTTP client is not being used by default, and only works if it is available on the classpath:

HTTP clients become somewhat pluggable with this change as well.
A specific HTTP client can be chosen via the `nessie.client-builder-name` configuration option:
* `HTTP` (default) chooses the default HTTP client
* `JavaHttp` forces the new Java HTTP client (default when running on Java 11 or newer)
* `URLConnection` forces the URLConnection based client (default when running on Java 8)
* `ApacheHttp` forces the Apache HTTP client

The configuration option `nessie.force-url-connection-client` is deprecated now, in favor of `nessie.client-builder-name=URLConnection`.