OAuth2 client config: "bark" with all check-errors (#8322)
The current validation of oauth2 options fails immediately for the first validation error. This can be more user friendly by throwing one exception with all violations, which is what this change does. It also replaces the "NPE-validation" for client-ID+secret from `OAuth2AuthenticatorConfig` to "IAE-validation" in `OAuth2ClientConfig`.

The change results in better validation error messages, which also include the config parameter key.