Site: exclude generated .md files from search index
The generated markdown files, meant to be only included in other pages, sadly still occur in the site search results as individual pages, which is not intended.

This change updates the site generation by adding a 'dont include in search' to the front-matter of the generated files, including "latest/generated".