Simpler workaround for #8390 (#8397)
This change effectively reverts the hacky approach from #8394 and uses a simpler approach.

Initially it looked that https://github.com/quarkusio/quarkus/issues/40236 could be fixed relatively soon, but handling MR-jars is a rabbit-hole. There might at some point be a solution/fix for MR-jars in Quarkus, but likely not soon.

So this change introduces a new artifact `:nessie-model-quarkus`, which is solely meant to be consumed by `:nessie-quarkus`. It is _not_ a MR file, so it retains the `jakarta.*` annotations, which prevents Quarkus/resteasy-reactive running into the MR-jar rabbit-hole.

`:nessie-model-quarkus` just builds a jar using the exact same sources and build file instructions as `:nessie-model`, so it can serve as a proper replacement for `:nessie-model` in `:nessie-quarkus`