OAuthClient: minor tweaks to the token exchange flow
This PR comes from investigation of #8331: in some
cases, Keycloak stops updating the expiration time
of returned access tokens. As a consequence, the
token eventually expires, and the client recovers
by re-fetching using the initial grant.

However, if in that tiny window the expired token
is used, that would result in an HTTP 401.

This PR also changes the requested token type to
access token. Keycloak accepts both types (access
or refresh). But since we only use token exchange
when refresh tokens are disabled, it doesn't make
sense to request a refresh token.