Add license reports and checks (#8497)
Dependencies, runtime dependencies including those that are distributed as part of a fat/uber jar, must use a "compatible" license. Checking those licenses manually is good, but some automatic checks are better.

This change adds a Gradle plugin to check licenses in the runtime configurations of the Nessie distributables (see below), and to produce license reports as HTML.

Nessie distributable are:
* Nessie Quarkus Server
* Nessie Quarkus CLI
* Nessie GC tool
* Nessie Content Generator
* Nessie Spark SQL extensions
* Nessie CLI/REPL

Checks and reports are run against the `runtimeClasspath` configurations of the above.