Persistence: properly handle timeout-ish exceptions
Databases can run into timeout situations, which should not happen, but do happen.

This change introduces a new `UnknownOperationResultException`, which signals that the database operation did not return a meaningful result and the outcome of the operation is unknown. Within the Nessie commit-retry loop this exception is handled via a "regular" commit-retry.