Split Release workflow job into multiple, retryable jobs
The Nessie release for 0.83 failed two times, we had to re-release as 0.83.1 and finally 0.83.2, beacuse the release-publish workflow is not retryable, mostly because multiple publications of Maven artifacts of the same versions are not possible.

This change splits the single workflow job into multiple:

```
Prepare -->--+
             |
             +-->-- Publish to Maven Central -->--+
             |                                    |
             +-->-- Publish OpenAPI ----------->--+
             |                                    |
             +-->-- Publish Docker images ----->--+
             |                                    |
             +-->-- Publish Helm Chart -------->--+
                                                  |
                                                  +-->-- GitHub Release notes
```

This change also moves the big script to create the GH release notes to a separate shell script.

Fixes #8581