OAuth2Client: refactor scopes (#8814)
This change refactors the API around OAuth scopes and turns this
attribute into a list of scopes, instead of a space-separated
string list. This makes the programmatic configuration slightly
more natural.

This change has no effect on configuration, which remains the same
(scopes must be passed in one single option as a space-separated
string list).