Use `search` in `resolv.conf` to resolve service names
The name of the Nessie/Quarkus management service is provided without a domain name, but the lookup needs to happen using the fully qualified domain name (FQDN). This change adds the ability to use the `search` list provided in `/etc/resolv.conf` in `AddressResolver`, via the utility class `ResolvConf`.

Verified locally with minikube with 1, 2 and 3 pods (scaling up).

Fixes #8829