Committing operations: cleanup metadata after failures
Ensures that objects/files written during a commit are cleaned up if either one of the object/file write operations or the Nessie commit operation fails.

Refactors `CatalogServiceImpl.MultiTableUpdate` to be a top-level type and adds a list of successfully written objects/files in a separate `exceptionally()` stage, that calls `ObjectIO.deleteObjects()` for the written objects/files.