Make address resolution more resilient (#8984)
This PR was initially only to resolve the flaky flaky `TestAddressResolver.resolveGoodName`. It turned out, that the flakiness was "just" a too short DNS query timeout of 250ms. which can easily be too low to resolve names from external DNS.

However, the fact that failed resolutions are recovered to an "empty list" can be problematic. So this PR also removes the "silent recovery" and propagates DNS resolution errors, which are then "handled" by `CacheInvalidationSender` by not using an incomplete result.