Export related objects
Export and import objects related to the repository, the references, the commits and contents.

The IDs of the related objects to be exported are retrieved via `TransferRelatedObjects` implementations, introduced by #9033. Those are exported and imported using the ability to (de)serialize "any" custom object, introduced by #9032.

This change also bumps the export version of the "contents export" from 1 to 3 to handle the related objects, including the necessary export changes.