Catalog: Improve warehouse health/readiness checks
The current health check for warehouses may not expose actable information, for example just some `java.util.NoSuchElementException`, which is not helpful.

This change logs the stack trace in the server logs, associated with a random error-ID, which is included in the health-check response, allowing admins to look up the stacktrace by that ID.

Also send the stack trace back to the client, if `nessie.server.send-stacktrace-to-client` is `true` (not recommended).