Ability to identify related objects for Nessie export (#9033)
Nessie export needs to be able to identify which objects are related to the repository, references, commits and contents that are being exported.

This service-loader based functionality allows Nessie export to identify those objects. Each relevant module provides an implementation of the `TransferRelatedObjects` interface, ideally in an artifact that has no "superfluous" dependencies, so only the code that is required to compute the `ObjId`s. This change includes two implementations of `TransferRelatedObjects`, one to provide the `ObjId` of the `UniqueIdObj` for each `Content` and one for Nessie Catalog.