Catalog: Fix assignment of nested fields in a table's initial schema (#9085)
The field IDs of nested structures/lists/maps are not properly (re)assigned during the creation of a table. This can lead to duplicate field IDs.

This change fixes the oversight and adds a test for this case.

Another bug fixed by this change it the calculation of the `last-column-id`, which needs to even respect IDs in collection types (maps + lists).