CLI: Add `REVERT CONTENT` command
Based on a user's idea, this new `REVERT CONTENT` command allows updating one or more individial content keys to a state as in another commit.

The example at hand is that the user's query engine performed a commit to Nessie, but removed the metadata object after that, leaving a Nessie commit w/ an invalid metadata location.