Initial Trino integration testing
This change adds a new `TrinoContainer` to test Nessie Catalog against Trino.

It would have been easier to just use `trino-jdbc`, but that clashes horribly with Quarkus, because `trino-jdbc` has OpenTelemetry instrumentation built in using shaded classes, but not all of them - and the opentelementry-semconv from Quarkus is incompatible with the one hardly required by `trino-jdbc`. To have "something", this change adds rudimental support to run queries, which is hopefully enough for testing. Beside, having to pull in okhttp isn't especially great, but there's no way around :(

The `ITTrinoS3` class uses a somewhat awkward mechanism to let Trino connect to Nessie, but it works using testcontainer's mechanism to expose host ports to containers.