Catalog: Enable time-travel and branch/tag selection for Iceberg REST in all cases
Currently on the functionality for time-travel and branch/tag selection in Nessie's Iceberg-REST is not fully functional. This change adjusts the code to handle branch/tag selection and time-travel functionality and adds tests for that. The syntax is the same as the one used in Nessie's Spark extensions.

Also add a clarifying chapter to the web site.