Catalog: Implementation for short-lived user-delegation SAS tokens (#9301)
Implements the production code for short-lived user-delegation SAS tokens. Sadly, Azurite doesn't work with this, so the code is effectively untested. In a follow-up it feels appropriate to cache the user-delegation-keys, which have a relatively long lifetime and are needed to generate the actual SAS tokens (only the latter are passed down to clients).

Fixes #8909