Explicitly prevent empty content keys
Empty content keys don't make any sense, but are theoretically possible. Empty content keys _could_ cause weird issues, so disallowing them. Also, empty content with an empty key is not be "addressable" at all in any tool/client, so the chance that any content with an empty key exists in the wild is 0.

Fixes #2648