Reference secrets by name, do not inject
Warehouses/object-stores require secrets (access-keys, tokens, etc). These secrets are currently injected into the object-store specific configuration holders, when the effective bucket configuration is calculated.

The configuration of warehouses/object-stores should be distinct from the definition of secrets. This requires to not have the secrets in a bucket configuration at any time, but instead _reference_ secrets by name. These names are then resolved to secrets.

This change replaces the `Secret` types in all bucket configuration types with a `String` as the _reference_ to the secret.

Also implicitly refactors most of `:nessie-catalog-secrets-api` module.