Disable Swagger-UI + eliminate "duplicate operation ID" warnings in Quarkus tests (#9340)
Disable the Swagger-UI in Quarkus, because the OpenAPI path "mismatches" between the interfaces in the `org.projectnessie.api.v1/2.http` packages and the implementations in the `org.projectnessie.services.rest` package. Those do either produce duplicate entries in the Swagger UI or, when only including the implementations via `mp.openapi.scan.include`, no entries. Also, the Swagger UI is rather a dev-mode thing and is only available on the Quarkus management port and therefore not publicly available. Also disables the Quarkus provided download of the OpenAPI yaml and replaced it with a download of the `openapi.yaml` from `:nessie-model`.