[DocTool] Allow nested config sections (#9370)
The current smallrye-config parts of the doc-tool can only handle "config sections" directly on the `@ConfigMapping` type, but adding `@ConfigItem(section = ...)` on any attribute in a nested type is not handled.

This change was primarily intended to just allow that, but turned out into a bigger refactoring and eventually a leaner code base. Generation of the smallrye docs is now a single recursive walk, which eliminates the problem of the (currenlty broken) handling of property prefixes (actually two variants: the "plain" and the "markdown" one) and section prefixes.

As a side effect, this change also fixes actually wrong property names in the reference docs. Examples:
* currently wrong: `nessie.catalog.warehouseDefaults.default-warehouse`, correct: `nessie.catalog.default-warehouse`
* currently wrong: `nessie.catalog.service.adls.transport....`, currect: `nessie.catalog.service.adls....`