Allow Nessie Web UI to authenticate (#9398)
Currently the Nessie Web UI cannot be used when authentication is enabled.

This change, which is mostly a set of Quarkus configurations, allows using the Nessie Web UI - however, there's no "login" or "logout" button.

The "trick" is to use a different Quarkus OIDC tenant (called `ui-app`), which is chosen by Quarkus for the static UI-path `/` (and it's single-page-app mapped paths).

Although it's possible to define a different "client" and even a different oauth-server, the Helm chart changes just duplicate the configured settings.

It's important to retain the existing behavior of the "service" (REST API) endpoints and to return `HTTP/401`, whereas the ui-app (a browser) requires `HTTP/302` to the authentication page. This works:
```
$ curl -v http://127.0.0.1:19120/api/v2/config
...
< HTTP/1.1 401 Unauthorized

$ curl -v http://127.0.0.1:19120/favicon.ico
< HTTP/1.1 200 OK

$ curl -v http://127.0.0.1:19120/
< HTTP/1.1 302 Found
< location: http://127.0.0.1:8080/realms/iceberg/protocol/openid-connect/auth?response_type=code&client_id=client1&scope=openid&redirect_uri=http%3A%2F%2F127.0.0.1%3A19120%2F&state=...
```

This change also allows anonymous access to the static Nessie OpenAPI yaml.