Derive `location` of new tables from parent namespaces, add some validations
Whether the `location` property/attribute can be set is delegated to access checks, therefore the currently existing restriction that the location cannot be set is relaxed w/ this change if no access control for it is in place.

This change also adds somewhat rudimentary checks whether the configured location is valid.

Fixes #9331