Persistable `LakehouseConfig`
This change allows to use the `LakehouseConfig` from the Quarkus configuration, which is the current behavior, and with change to alternatively persist the `LakehouseConfig`.

For this to work, the `LakehouseConfig` object tree needs to be injectable via the CDI request-scope, so some refactoring to make this work is included in this change.

Some options, which were still in the `LakehouseConfig`, had been moved into two object types `S3StsCache` (the old `S3Sts`, but moved out of the LC tree) and `GcsConfig`.

Retrieving and updating the `LakehouseConfig` object is implemented via `LakehouseConfigManagement` (plus its implementation).