ReferenceLogic: parameterized purge of the commit log of a `Reference`
Just implements the functionality to "shorten" the commit log history of a `Reference`. No API yet and this should still be optimized to reduce heap pressure by memoizing only `ObjId`s instead of `CommitObj`s.

Exposing this functionality is not yet adviseable, because it would be meaningles w/o #9688 in place and exposed as a functionality as well.

Fixes #9733