Testing: disable looking up GCP credentials (#9900)
This is a follow-up of #9866 (and it's revert #9899). The stack traces indicate that the global-tracer field init race happens when `BigTableBackendBuilder` tries to fetch the GCP credentials via `Instance<CredentialsProvider>`, which eventually lets the Google client try to find the default app-credentials, which then implicitly initializes OpenCensus->OpenTelemetry.

The workaround here is to add a new undocumented configuration option to disable GCP credentials lookup, which is also not necessary when using the BT emulator.