Add admin tool command to cut commit log at a certain point. (#10048)
The tool is intended for advanced use cases when commit log
needs to be shortened.

Cuts across branch points are not detected, but may lead
to errors. This is left to the end user to avoid.

Some child commits of the new history root are rewritten to
prevent their tails from referencing deeper parents.