Remove use of ScyllaDB in Nessie
ScyllaDB has changed its license a few days ago in a way that makes the use of ScyllaDB in Nessie, even just for tests, potentially problematic. Therefore Nessie will no longer test against ScyllaDB. The relevant production code as of Dec 27 2024 remains unchanged, but also unmaintained. The mention of ScyllaDB on the docs for the next release has been removed.