Adopt multi-env-test-engine for >= 5.12 and < 5.12
JUnit 5.12 added the new type `OutputDirectoryProvider` and a new parameter to the (internal) class `org.junit.jupiter.engine.config.DefaultJupiterConfiguration`. While updating the code for 5.12 is straight forward, we still have to support both 5.12 and < 5.12, because Quarkus has an enforced dependency to JUnit 5.10.x, even in the upcoming Quarkus 3.19 release. Therefore a new `JUnitCompat` class has been added that can deal with both scenarios. Since it's impossible to do a `super(...)` call in a constructor using Java reflection, `MultiEnvDelegatingJupiterConfiguration` has been added as well to satisfy that use case.