Remove links to swaggerhub
Nessie release publications fail for quite some time now due to `Error: Exceeded access limits for projectnessie.` in the `publish-openapi` release workflow job.

This change updates the release-workflow to not fail at all during swaggerhub api calls (hoping it may continue to work at some point), but remove the links to swaggerhub on the web site and in-tree docs.