Migrate from VLSI Jandex plugin to Kordamp Jandex plugin
One difference between both plugins is that the Kordamp generates a Jandex index only for the "main" jar, while the VLSI plugin creates a Jandex index for every jar. This differences leads to a removal of some build script code.

This PR also bumps the Jandex dependency version to the latest, but ensures that the Jandex index version (11) from Jandex 3.1.8 is being used.

This change solves the Quarkus 3.26.3 upgrade issue in #11265

Quarkus 3.26.3 also contains a fix for the boilerpolate `dependsOn`s.
The code for those has been moved into `nessie-conventions-quarkus`. Having the `.dependsOn`s there doesn't work for projects that don't use the Quarkus plugin, so this change also renames nessie-conventions-client to nessie-conventions-java11 and nessie-conventions-server to nessie-conventions-java21 (-client and -server were both targeting Java 11) and leaves nessie-conventions-quarkus exclusively for use w/ the Quarkus plugin.